6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int main(void)
 	float floating;
 	char character;
 
-	printf("size of int: %lu byte(s)\n", sizeof(integer));
-	printf("size of double: %lu byte(s)\n", sizeof(dou));
-	printf("size of float: %lu byte(s)\n", sizeof(floating));
-	printf("size of char: %lu byte(s)", sizeof(character));
+	printf("size of int: %lu byte(s)\n", (unsigned long)sizeof(integer));
+	printf("size of double: %lu byte(s)\n", (unsigned long)sizeof(dou));
+	printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(floating));
+	printf("size of char: %lu byte(s)", (unsigned long)sizeof(character));
 
 	return (0);
 }
